Move replay code back to replayBrowser
