+++ src/common/replayBrowser/stats.worker.js
